getting started edits
