--- docs/components/ExampleBarChart.vue
@@ -1,63 +1,0 @@
-<script>
-
-import * as d3 from "d3";
-
-export default {
-  async mounted() {
-    const alphabet = await d3.csv("https://static.observableusercontent.com/files/75d52e6c3130b1cae83cda89305e17b50f33e7420ef205587a135e8562bcfd22e483cf4fa2fb5df6dff66f9c5d19740be1cfaf47406286e2eb6574b49ffc685d?response-content-disposition=attachment%3Bfilename*%3DUTF-8%27%27alphabet.csv");
-
-    // Declare the chart dimensions and margins.
-    const width = 640;
-    const height = 400;
-    const marginTop = 30;
-    const marginRight = 0;
-    const marginBottom = 30;
-    const marginLeft = 40;
-
-    // Declare the x (horizontal position) scale.
-    const x = d3.scaleBand()
-        .domain(alphabet.map((d) => d.letter))
-        .range([marginLeft, width - marginRight])
-        .padding(0.1);
-
-    // Declare the y (vertical position) scale.
-    const y = d3.scaleLinear()
-        .domain([0, d3.max(alphabet, (d) => d.frequency)])
-        .range([height - marginBottom, marginTop]);
-
-    // Create the SVG container.
-    const svg = d3.create("svg")
-        .attr("width", width)
-        .attr("height", height);
-
-    // Add a rect for each bar.
-    svg.append("g")
-        .attr("fill", "currentColor")
-      .selectAll()
-      .data(alphabet)
-      .join("rect")
-        .attr("x", (d) => x(d.letter))
-        .attr("y", (d) => y(d.frequency))
-        .attr("height", (d) => y(0) - y(d.frequency))
-        .attr("width", x.bandwidth());
-
-    // Add the x-axis.
-    svg.append("g")
-        .attr("transform", `translate(0,${height - marginBottom})`)
-        .call(d3.axisBottom(x).tickSizeOuter(0));
-
-    // Add the y-axis.
-    svg.append("g")
-        .attr("transform", `translate(${marginLeft},0)`)
-        .call(d3.axisLeft(y));
-
-    // Append the SVG element.
-    this.$el.append(svg.node());
-  }
-}
-
-</script>
-
-<template>
-  <div></div>
-</template>
